Refactor: Rename model storage to vector storage
