+++ src/mem_agent/__init__.py
@@ -23,7 +23,11 @@ Architecture (SOLID principles):
 
 Storage Types:
   - "json": Fast, lightweight, no ML dependencies (default)
-  - "model": Semantic search using AI model, requires transformers/sentence-transformers
+  - "vector": Semantic search using embeddings, requires transformers/sentence-transformers
+
+Note:
+  This module provides the storage backend for MCP Memory tool.
+  "mem-agent" refers to a future LLM-based assistant implementation (not yet implemented).
 
 Settings are in config.settings module (MEM_AGENT_STORAGE_TYPE).
 
@@ -34,7 +38,7 @@ Installation:
 # Core interfaces and implementations
 from .base import BaseMemoryStorage
 from .json_storage import JsonMemoryStorage
-from .model_storage import VectorBasedMemoryStorage
+from .vector_storage import VectorBasedMemoryStorage
 from .factory import MemoryStorageFactory, create_memory_storage
 
 # Legacy compatibility
